Add NEKOXAU token info and logo
This PR adds token information and logo for NEKOXAU, an ERC-20 meme coin focused on community, patience, and digital legacy.

Although NEKOXAU is still in its early stage, we are gradually growing the holder base and liquidity transparently. We are submitting this PR in preparation for future visibility in Trust Wallet and to demonstrate our commitment to the project.

Contract Address: 0xbf824DAFA61F7c1FEc739b4914a7fF867284dcAd
Official Links:
- Website: https://nekoxau.carrd.co
- Twitter: https://twitter.com/abyssstride
- Telegram: https://t.me/NEKOXAUREAL

Thank you for reviewing this submission.